Preliminary bundling support
