+++ README.md
@@ -1,11 +1,17 @@
+<center>
+
 # procedure.js 🔗
 The simple RPC framework for Node.js.
 
-[![npm package version](https://img.shields.io/npm/v/@procedure-rpc/procedure.js.svg)](https://npmjs.org/package/@procedure-rpc/procedure.js "View procedure.js on npm") [![npm package downloads](https://img.shields.io/npm/dw/@procedure-rpc/procedure.js.svg)](https://npmjs.org/package/@procedure-rpc/procedure.js "View procedure.js on npm") [![API docs](https://img.shields.io/badge/docs-v0.6.1-informational.svg)](https://procedure-rpc.github.io/procedure.js "Read the documentation on Github Pages") [![Code coverage](https://img.shields.io/codecov/c/github/procedure-rpc/procedure.js?label=code%20coverage)](https://codecov.io/gh/procedure-rpc/procedure.js "View code coverage on Codecov") [![code quality](https://img.shields.io/codefactor/grade/github/procedure-rpc/procedure.js.svg)](https://www.codefactor.io/repository/github/procedure-rpc/procedure.js "Check code quality on CodeFactor") <!-- [![Coverity Scan build status](https://img.shields.io/coverity/scan/.svg)](https://scan.coverity.com/projects/procedure-rpc-procedure.js "View build status on Coverity Scan") --> <!-- ![nyc code coverage](https://img.shields.io/nycrc/procedure-rpc/procedure.js.svg)] -->
+[![npm package version](https://img.shields.io/npm/v/@procedure-rpc/procedure.js.svg)](https://npmjs.org/package/@procedure-rpc/procedure.js "View procedure.js on npm") [![npm package downloads](https://img.shields.io/npm/dw/@procedure-rpc/procedure.js.svg)](https://npmjs.org/package/@procedure-rpc/procedure.js "View procedure.js on npm") [![API docs](https://img.shields.io/badge/docs-v0.7.0-informational.svg)](https://procedure-rpc.github.io/procedure.js "Read the documentation on Github Pages")
+
+[![Code coverage](https://img.shields.io/codecov/c/github/procedure-rpc/procedure.js?label=code%20coverage)](https://codecov.io/gh/procedure-rpc/procedure.js "View code coverage on Codecov") [![code quality](https://img.shields.io/codefactor/grade/github/procedure-rpc/procedure.js.svg)](https://www.codefactor.io/repository/github/procedure-rpc/procedure.js "Check code quality on CodeFactor") [![minzip](https://badgen.net/bundlephobia/minzip/@procedure-rpc/procedure.js)](https://bundlephobia.com/package/@procedure-rpc/procedure.js "View procedure.js on Bundlephobia") [![tree-shaking](https://badgen.net/bundlephobia/tree-shaking/@procedure-rpc/procedure.js)](https://bundlephobia.com/package/@procedure-rpc/procedure.js "View procedure.js on Bundlephobia")
 
 [![npm test](https://github.com/Procedure-RPC/procedure.js/actions/workflows/npm-test.yml/badge.svg)](https://github.com/Procedure-RPC/procedure.js/actions/workflows/npm-test.yml "View npm test on GitHub Actions") [![publish package](https://github.com/Procedure-RPC/procedure.js/actions/workflows/publish-package.yml/badge.svg)](https://github.com/Procedure-RPC/procedure.js/actions/workflows/publish-package.yml "View publish package on GitHub Actions") [![publish docs](https://github.com/Procedure-RPC/procedure.js/actions/workflows/publish-docs.yml/badge.svg)](https://github.com/Procedure-RPC/procedure.js/actions/workflows/publish-docs.yml "View publish docks on GitHub Actions")
 
 [![GitHub stars](https://img.shields.io/github/stars/procedure-rpc/procedure.js.svg?style=social)](https://github.com/procedure-rpc/procedure.js "Star procedure.js on GitHub") [![Twitter Follow](https://img.shields.io/twitter/follow/toebean__.svg?style=social)](https://twitter.com/toebean__ "Follow @toebean__ on Twitter") [![GitHub Sponsors donation button](https://img.shields.io/badge/github-sponsor-yellow.svg)](https://github.com/sponsors/toebeann "Sponsor procedure.js on GitHub") [![PayPal donation button](https://img.shields.io/badge/paypal-donate-yellow.svg)](https://paypal.me/tobeyblaber "Donate to procedure.js with PayPal")
+
+</center>
 
 ## Description
 A lightweight alternative to the boilerplate-heavy gRPC, or spinning up a HTTP server and REST API. Procedure links your independent applications and services with as little code as possible, so that you can just focus on building your app!
