+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "@procedure-rpc/procedure.js",
-  "version": "0.6.1",
+  "version": "0.7.0",
   "title": "procedure.js",
   "description": "The simple RPC framework for Node.js.",
-  "main": "./dist/index",
-  "types": "./types/index.d.ts",
+  "main": "./dist/index.cjs.js",
+  "module": "./dist/index.esm.js",
+  "types": "./types/index.esm.d.ts",
   "author": "Tobey Blaber (https://github.com/toebeann)",
   "homepage": "https://procedure-rpc.github.io/procedure.js",
   "repository": "github:procedure-rpc/procedure.js",
@@ -15,7 +16,7 @@
   "license": "MIT",
   "sideEffects": false,
   "scripts": {
-    "build": "tsc",
+    "build": "rollup -c",
     "lint": "eslint . --ext .ts",
     "test": "mocha -r ts-node/register -r source-map-support/register ./test/**/*.ts",
     "test:coverage": "npm run build & c8 npm test",
@@ -23,11 +24,13 @@
   },
   "dependencies": {
     "@msgpack/msgpack": "^2.7.2",
-    "@toebean/signals": "^2.1.0",
+    "@toebean/signals": "^2.2.0",
     "nanomsg": "^4.2.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "@rollup/plugin-node-resolve": "^13.3.0",
     "@tsconfig/node16": "^1.0.3",
     "@types/chai": "^4.3.1",
     "@types/chai-as-promised": "^7.1.5",
@@ -44,8 +47,13 @@
     "chai-spies": "^1.0.0",
     "eslint": "^8.20.0",
     "mocha": "^10.0.0",
+    "rollup": "^2.78.1",
+    "rollup-plugin-polyfill-node": "^0.10.2",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-ts": "^3.0.2",
     "source-map-support": "^0.5.21",
     "ts-node": "^10.9.1",
+    "tslib": "^2.4.0",
     "typed-emitter": "^2.1.0",
     "typedoc": "^0.23.10",
     "typescript": "^4.7.4"
