+++ tsconfig.json
@@ -7,7 +7,8 @@
     "declarationMap": true,
     "declarationDir": "./types",
     "sourceMap": true,
-    "strictNullChecks": true
+    "strictNullChecks": true,
+    "isolatedModules": true
   },
   "include": [
     "./src"
